backend/requirements/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
pypandoc==1.12
python-docx==1.1.0
markdown==3.5.1
asyncio
threading
uuid
datetime
tempfile
shutil
os
sys
collections

# Dependencies for the report generator
pydantic==2.5.0
langchain==0.1.0
langchain-core==0.1.0
tavily-python==0.3.0
python-dotenv==1.0.0
langgraph==0.0.20 